Refactor group menu functionality and update DTOs
- Updated `IGroupMenuService` and `IGroupMenuRepo` to use `GroupMenuDTO` instead of `SearchGroupMenuDTO` and added `GetAllMenusByGroup` method.
- Renamed `SearchGroupMenuDTO` to `GroupMenuDTO` and adjusted boolean properties to be non-nullable.
- Modified `CreateGroupMenuCommand` to set `Access` default to `true`.
- Updated handlers to ensure consistency with the new DTO.
- Implemented `GetAllMenusByGroup` in `GroupMenuRepo` and `GroupMenuService`.
- Added new methods in `CompanyCapitalRepo` and `CompanyShareHolderRepo`.
- Introduced `GetAllMenusByGroupHandler` and `GetAllMenusByGroupQuery` for handling menu retrieval.
- Updated `GroupMenuController` to include a new endpoint for fetching menus by group.
- Adjusted CORS policy in `Program.cs` and modified connection string in `appsettings.Development.json`.
- Added necessary using directives across various files.

+++ Nexus.LAS.Application/DTOs/CompanyDTOs/GetCompaniesDto.cs
@@ -24,7 +24,7 @@ public class GetCompanyDto
     public bool? CciExpiryActiveReminder { get; set; }
     public int? PlaceOfRegistrationMainIdn { get; set; }
     public int? PlaceOfRegistrationSubIdn { get; set; }
-    public decimal? CapitalAmount { get; set; }
+    public double? CapitalAmount { get; set; }
     public long? TotalShares { get; set; }
     public int? NumberOfPartners { get; set; }
     public DateTime? UpdateDate { get; set; }
